% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KVA.R
\name{KVA}
\alias{KVA}
\title{Knowledge Value Added - Training Time}
\usage{
KVA(
  no_emp,
  learn_time,
  pctauto,
  times_perf_year,
  avg_time_complete,
  wage,
  revenue
)
}
\arguments{
\item{no_emp}{Number of employees}

\item{learn_time}{Learning time by task}

\item{pctauto}{Percent of task automated (enter as Decimal)}

\item{times_perf_year}{Times a task is performed in a year}

\item{avg_time_complete}{Average time to completion of a task}

\item{wage}{Wages paid by task}

\item{revenue}{Revenue by task}
}
\value{
a 3 item list containing [[1]]: Revenue per Unit of knowledge, [[2]]: Return on Knowledge by task, and [[3]]: a dataframe compiling all calculations and input data
}
\description{
Using Training time as a unit of input, generates a 3 item list containing [[1]]: Revenue per Unit of knowledge, [[2]]: Return on Knowledge by task, and [[3]]: a dataframe compiling all calculations and input data
}
\examples{
KVA(c(1, 15, 1), c(11, 47, 16), c(0.9, 0.14, 0.6), c(690, 6500, 230), c(0.5, .33, 1), c(9.5, 12, 20), c(4916.25,  579150, 6900))

}
